+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Changes to the plugin.
 - Ability to turn on showing text for only parts of a randomtest run using `*comment text off` and `*comment text on`.
 - Count of the Randomtest test now shows on the status bar while the test is running.
 
+### Changed
+
+- Choices in the outline now include a summary of the first choice.
+
 ## [2.5.1] - 2022-01-23
 
 Point release to fix an issue with the Visual Studio Marketplace.